src/main/java/ma/m3achaba/plantes/handler/GlobalExcepHandler.java
package ma.m3achaba.plantes.handler;

import lombok.extern.slf4j.Slf4j;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;

import java.time.LocalDateTime;

import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;

@RestControllerAdvice
@Slf4j
public class GlobalExcepHandler {
    @ExceptionHandler(Exception.class)
    public ResponseEntity<ExceptionResponse> handleException(Exception exp) {
        log.error(exp.getMessage(), exp);
        return ResponseEntity
                .status(INTERNAL_SERVER_ERROR)
                .body(
                        ExceptionResponse.builder()
                                .details("Erreur interne, veuillez contacter l'administrateur")
                                .message(exp.getMessage())
                                .code(INTERNAL_SERVER_ERROR.value())
                                .date(LocalDateTime.now())
                                .build()
                );
    }
}
